+++ src/main/c/cext/printf.c
@@ -56,13 +56,17 @@ VALUE rb_enc_vsprintf(rb_encoding *enc, const char *format, va_list args) {
   VALUE types = RUBY_CEXT_INVOKE("rb_tr_sprintf_types", rubyFormat);
   VALUE rubyArgs = rb_tr_get_sprintf_args(args, types);
 
-  return rb_str_conv_enc(rb_tr_wrap(
-                    polyglot_invoke(
-                                    RUBY_CEXT,
-                                    "rb_tr_sprintf",
-                                    rb_tr_unwrap(rubyFormat),
-                                    rb_tr_vsprintf_new_cstr,
-                                    rb_tr_unwrap(rubyArgs))), NULL, enc);
+  VALUE string = rb_tr_wrap(
+    polyglot_invoke(
+      RUBY_CEXT,
+      "rb_tr_sprintf",
+      rb_tr_unwrap(rubyFormat),
+      rb_tr_vsprintf_new_cstr,
+      rb_tr_unwrap(rubyArgs)
+    )
+  );
+
+  return rb_enc_associate(string, enc);
 }
 
 VALUE rb_enc_sprintf(rb_encoding *enc, const char *format, ...) {