+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Bug fixes:
 Compatibility:
 
 * Add `Exception#detailed_message` method (#3257, @andrykonchin).
+* Fix `rb_enc_vsprintf` and force String encoding instead of converting it (@andrykonchin).
 
 Performance: